# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/graph/stats:
    get:
      operationId: GraphController_stats
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphStatsDto'
      security:
      - bearer: []
      summary: Graph statistics — counts per node type
      tags:
      - Graph
  /api/v1/graph/nodes:
    get:
      operationId: GraphController_listNodes
      parameters:
      - name: type
        required: false
        in: query
        description: Filter by node type
        schema:
          example: Screen
          type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphNodesResponse'
        '401':
          description: Unauthorized
      security:
      - bearer: []
      summary: List graph nodes, optionally filtered by type
      tags:
      - Graph
  /api/v1/graph/nodes/{id}:
    get:
      operationId: GraphController_getNode
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Node found
        '404':
          description: Node not found
      security:
      - bearer: []
      summary: Get a single graph node by ID
      tags:
      - Graph
  /api/v1/graph/edges:
    get:
      operationId: GraphController_listEdges
      parameters:
      - name: type
        required: false
        in: query
        description: Filter by edge type
        schema:
          example: DEPENDS_ON
          type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GraphEdgeDto'
      security:
      - bearer: []
      summary: List graph edges, optionally filtered by type
      tags:
      - Graph
  /api/v1/graph/edges/{id}:
    get:
      operationId: GraphController_getEdge
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Edge found
        '404':
          description: Edge not found
      security:
      - bearer: []
      summary: Get a single edge by ID
      tags:
      - Graph
  /api/v1/graph/types:
    get:
      operationId: GraphController_nodeTypes
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NodeTypeCount'
      security:
      - bearer: []
      summary: List all node types with counts
      tags:
      - Graph
  /api/v1/impact/analyze:
    post:
      operationId: ImpactController_analyze
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImpactAnalyzeRequest'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImpactAnalyzeResponse'
        '400':
          description: changeRequest.title is required
      security:
      - bearer: []
      summary: Analyze impact of a change request on services, screens, and integrations
      tags:
      - Impact
  /api/v1/impact/jobs:
    post:
      operationId: ImpactController_createJob
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '201':
          description: Job created
        '400':
          description: title is required
      security:
      - bearer: []
      summary: Create a new async impact analysis job
      tags:
      - Impact
    get:
      operationId: JobsController_list
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'
      security:
      - bearer: []
      summary: List all async impact analysis jobs
      tags:
      - Jobs
  /api/v1/impact/search:
    get:
      operationId: ImpactController_search
      parameters:
      - name: q
        required: true
        in: query
        description: Search query (min 2 chars)
        schema:
          example: ترقيات
          type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Query too short
      security:
      - bearer: []
      summary: Search across all graph nodes by name or description
      tags:
      - Impact
  /api/v1/impact/jobs/{jobId}:
    get:
      operationId: JobsController_get
      parameters:
      - name: jobId
        required: true
        in: path
        schema:
          example: job-abc123
          type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDetailResponse'
        '404':
          description: Job not found
      security:
      - bearer: []
      summary: Get job status and details
      tags:
      - Jobs
  /api/v1/impact/jobs/{jobId}/result:
    get:
      operationId: JobsController_getResult
      parameters:
      - name: jobId
        required: true
        in: path
        schema:
          example: job-abc123
          type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResultResponse'
        '404':
          description: Job not found
      security:
      - bearer: []
      summary: Get the result of a completed job
      tags:
      - Jobs
...
  /api/v1/external-systems:
    get:
      operationId: ExternalSystemsController_list
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalSystemListResponse'
      security:
      - bearer: []
      summary: List all external systems with integration counts
      tags:
      - External Systems
  /api/v1/external-systems/{id}:
    get:
      operationId: ExternalSystemsController_getById
      parameters:
      - name: id
        required: true
        in: path
        description: 'External system ID (with or without ext: prefix)'
        schema:
          example: ext:nic
          type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalSystemDetailResponse'
        '404':
          description: External system not found
      security:
      - bearer: []
      summary: Get external system detail with connected services and integrations
      tags:
      - External Systems
info:
  title: MASAR Impact Analysis API
  description: API for impact analysis of MASAR government services. Includes service
    graph traversal, async job processing, and Mermaid diagram generation.
  version: 2.0.0
  contact: {}
...
servers:
- url: https://myapi.businessanalystcrew.org
  description: Production
- url: http://localhost:3000
  description: Local
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: API token from system administrator
  schemas:
    GraphStatsDto:
      type: object
      properties:
        total_nodes:
          type: number
          example: 364
        total_edges:
          type: number
          example: 406
        by_type:
          type: object
          example:
            Service: 25
            Screen: 64
      required:
      - total_nodes
      - total_edges
      - by_type
    GraphNodeRef:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        label:
          type: string
      required:
      - id
      - type
      - label
    GraphNodesResponse:
      type: object
      properties:
        count:
          type: number
          example: 64
        type:
          type: object
          example: Screen
          nullable: true
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/GraphNodeRef'
      required:
      - count
      - type
      - nodes
    GraphEdgeDto:
      type: object
      properties:
        id:
          type: string
          example: edge:svc:wafed:has_screen:screen:wafed:contract-submit
        source_id:
          type: string
          example: svc:wafed
        target_id:
          type: string
          example: screen:wafed:contract-submit
        type:
          type: string
          example: HAS_SCREEN
        evidence_ids:
          type: array
          items:
            type: string
        impact_strength:
          type: string
          example: high
        impact_category:
          type: string
          example: functional
        reason:
          type: string
        confidence:
          type: number
          example: 0.9
      required:
      - id
      - source_id
      - target_id
      - type
      - evidence_ids
      - impact_strength
      - impact_category
      - confidence
    NodeTypeCount:
      type: object
      properties:
        type:
          type: string
          example: Service
        count:
          type: number
          example: 25
      required:
      - type
      - count
    ChangeRequestInput:
      type: object
      properties:
        title:
          type: string
          example: تعديل شاشة إنهاء التعاقد
        description:
          type: string
          example: إضافة إدخال يدوي لفئة الوظيفة والمسمى الوظيفي
        targetService:
          type: string
          example: svc:wafed
      required:
      - title
    ImpactAnalyzeRequest:
      type: object
      properties:
        changeRequest:
          $ref: '#/components/schemas/ChangeRequestInput'
      required:
      - changeRequest
    ImpactSummary:
      type: object
      properties:
        totalServices:
          type: number
          example: 3
        highImpact:
          type: number
          example: 1
        mediumImpact:
          type: number
          example: 2
        risks:
          type: number
          example: 0
      required:
      - totalServices
      - highImpact
      - mediumImpact
      - risks
    ImpactRisk:
      type: object
      properties:
        type:
          type: string
          example: integration
        severity:
          type: string
          example: high
        description:
          type: string
          example: JICS dependency may cause cascading failures
        recommendation:
          type: string
          example: Review integration contracts before change
      required:
      - type
      - severity
      - description
      - recommendation
    ImpactFinding:
      type: object
      properties:
        type:
          type: string
          example: info
        message:
          type: string
          example: Screen fields map to external system sarf
      required:
      - type
      - message
    ImpactAnalyzeResponse:
      type: object
      properties:
        requestId:
          type: string
          example: REQ-ABC123
        status:
          type: string
          example: completed
        timestamp:
          type: string
        summary:
          $ref: '#/components/schemas/ImpactSummary'
        impacts:
          type: array
          items:
            type: object
        risks:
          type: array
          items:
            $ref: '#/components/schemas/ImpactRisk'
        recommendedActions:
          type: array
          items:
            type: string
        findings:
          type: array
          items:
            $ref: '#/components/schemas/ImpactFinding'
      required:
      - requestId
      - status
      - timestamp
      - summary
      - impacts
      - risks
      - recommendedActions
      - findings
    CreateJobRequest:
      type: object
      properties:
        title:
          type: string
          example: تحليل تأثير تعديل وافد
        description:
          type: string
          example: تغيير شاشة إنهاء التعاقد
        targetService:
          type: string
          example: svc:wafed
      required:
      - title
    SearchResult:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        label:
          type: string
      required:
      - id
      - type
      - label
    SearchResponse:
      type: object
      properties:
        query:
          type: string
          example: ترقيات
        count:
          type: number
          example: 5
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
      required:
      - query
      - count
      - results
    JobDto:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
          example: تحليل تأثير تعديل وافد
        description:
          type: string
        targetService:
          type: string
        status:
          type: string
          example: completed
          enum:
          - queued
          - running
          - completed
          - failed
        progress:
          type: number
          example: 100
        createdAt:
          type: string
        startedAt:
          type: string
        completedAt:
          type: string
        error:
          type: string
        resultId:
          type: string
      required:
      - id
      - title
      - status
      - progress
      - createdAt
    JobListResponse:
      type: object
      properties:
        count:
          type: number
          example: 2
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/JobDto'
      required:
      - count
      - jobs
    JobDetailResponse:
      type: object
      properties:
        job:
          $ref: '#/components/schemas/JobDto'
      required:
      - job
    JobResultResponse:
      type: object
      properties:
        result:
          type: object
      required:
      - result
...
    ExternalSystemListItem:
      type: object
      properties:
        id:
          type: string
          example: ext:nic
        name_ar:
          type: string
          example: مركز المعلومات الوطني
        name_en:
          type: string
        connected_services:
          example:
          - svc:wafed
          - svc:bayan
          type: array
          items:
            type: string
        integration_count:
          type: number
          example: 15
      required:
      - id
      - name_ar
      - connected_services
      - integration_count
    ExternalSystemListResponse:
      type: object
      properties:
        count:
          type: number
          example: 25
        external_systems:
          type: array
          items:
            $ref: '#/components/schemas/ExternalSystemListItem'
      required:
      - count
      - external_systems
    IntegrationPointRef:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: تكامل مع مركز المعلومات الوطني
      required:
      - id
      - name
    IntegrationServiceRef:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: وافد
      required:
      - id
      - name
    IntegrationDetail:
      type: object
      properties:
        integration_point:
          $ref: '#/components/schemas/IntegrationPointRef'
        services:
          type: array
          items:
            $ref: '#/components/schemas/IntegrationServiceRef'
      required:
      - integration_point
      - services
    ExternalSystemDetailResponse:
      type: object
      properties:
        external_system:
          type: object
        integrations:
          type: array
          items:
            $ref: '#/components/schemas/IntegrationDetail'
      required:
      - external_system
      - integrations