# src/jobs/jobs.dto.ts
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';

export class JobDto {
  @ApiProperty() id: string;
  @ApiProperty({ example: 'تحليل تأثير تعديل وافد' }) title: string;
  @ApiPropertyOptional() description?: string;
  @ApiPropertyOptional() targetService?: string;
  @ApiProperty({ example: 'completed', enum: ['queued', 'running', 'completed', 'failed'] }) status: string;
  @ApiProperty({ example: 100 }) progress: number;
  @ApiProperty() createdAt: string;
  @ApiPropertyOptional() startedAt?: string;
  @ApiPropertyOptional() completedAt?: string;
  @ApiPropertyOptional() error?: string;
  @ApiPropertyOptional() resultId?: string;
}

export class JobListResponse {
  @ApiProperty({ example: 2 }) count: number;
  @ApiProperty({ type: [JobDto] }) jobs: JobDto[];
}

export class JobDetailResponse {
  @ApiProperty() job: JobDto;
}

export class JobResultResponse {
  @ApiProperty() result: Record<string, unknown>;
}